requirements.txt
matplotlib==3.5.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
Pillow==8.4.0
tqdm
numpy